Added Sensor interface

+++ SFMLSpaceGame/UI/ShipEditorWindow.cpp
@@ -12,8 +12,8 @@ void ShipEditorWindow::Update()
 {
 	MAKE_HIERARCHY(
 		INIT_AND_REFRESH(UIExpandingBackground, m_ids[0],
-			ON_INIT(UI_BACKGROUND_METAL_NINE, sf::Vector2f(500, 400))),
+			ON_INIT(UI_BACKGROUND_METAL_NINE, sf::Vector2i(500, 400))),
 		INIT_AND_REFRESH(UIExpandingBackground, m_ids[1],
-			ON_INIT(UI_BACKGROUND_METAL_NINE, sf::Vector2f(100, 100), UITransform(50, 50))));
+			ON_INIT(UI_BACKGROUND_METAL_NINE, sf::Vector2i(100, 100), UITransform(50, 50))));
 	
 }
